+++ CHANGES.md
@@ -4,6 +4,8 @@ Unreleased
 * Add the `[:alpha:]` character class in `Re.Perl` (#169)
 * Double asterisk (`**`) in `Re.Glob` (#172)
   Like `*` but also match `/` characters when `pathname` is set.
+* Double asterisk should match 0 or more directories unless in trailing
+  position. (#192, fixes #185)
 
 1.9.0 (05-Apr-2019)
 -------------------